atualização das páginas e controller de curso

+++ ps13si_new/src/main/java/br/com/elit/ps13si/repository/CursosRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.elit.ps13si.model.CursoModel;
 
-public interface CursosRepository extends JpaRepository<CursoModel, Integer> {
+public interface CursosRepository extends JpaRepository<CursoModel, Long> {
 
 }
